Move images to `images/` folder
