+++ src/views/Home/components/About/index.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import "./index.scss"
-import { ReactComponent as BackgroundLineMobile } from "images/icons/about-background-line-mobile.svg"
-import { ReactComponent as BackgroundLineDesktop } from "images/icons/about-background-line-desktop.svg"
+import { ReactComponent as BackgroundLineMobile } from "images/about/about-background-line-mobile.svg"
+import { ReactComponent as BackgroundLineDesktop } from "images/about/about-background-line-desktop.svg"
 import { AboutProps } from "Props/props"
 import AboutInformation from "Props/about/about"
 import AboutBox from "./AboutBox/AboutBox"
