+++ src/views/Home/components/Landing/index.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react"
-import { ReactComponent as Visual } from "images/landing/visual-desktop.svg"
-import { ReactComponent as VisualMobile } from "images/landing/visual-mobile.svg"
-import { ReactComponent as Hills } from "images/landing/hills.svg"
-import { ReactComponent as HillsMobile } from "images/landing/hills-mobile.svg"
+import { ReactComponent as Visual } from "images/landing/landing-visual-desktop.svg"
+import { ReactComponent as VisualMobile } from "images/landing/landing-visual-mobile.svg"
+import { ReactComponent as Hills } from "images/landing/landing-hills.svg"
+import { ReactComponent as HillsMobile } from "images/landing/landing-hills-mobile.svg"
 import "./index.scss"
 import Button from "components/Button/Button"
 import EmailSubscription from "components/EmailSubscription"
